[archi] reorganise index file
