+++ assembly/std/contract/index.ts
@@ -1,0 +1,2 @@
+export * from './calls';
+export * from './bytecode';
