+++ assembly/std/index.ts
@@ -41,610 +41,12 @@
  *
  */
 
-import { env } from '../env/index';
 import { Address } from './address';
 import * as Storage from './storage';
 import * as Context from './context';
-import { Args } from '@massalabs/as-types';
 
 export { Address, Storage, Context };
-
-/**
- * Logs a string message in the massa-node logs.
- *
- * @param message - The string message to be logged in the node.
- *
- */
-export function print(message: string): void {
-  env.print(message);
-}
-
-/**
- * Calls a function of a smart contract deployed at a given address.
- *
- * @remarks
- * The serialization of arguments must be handled by the caller and the callee.
- *
- * @param at - The address of the contract where the function will be executed.
- * @param functionName - The name of the function to be called in the contract.
- * @param args - The arguments of the function being called (type: Args).
- * @param coins - The amount of coins to pass if it is a payable function (minimum value is 0).
- *
- * @returns The return value of the executed function, serialized as a 'StaticArray<u8>'.
- *
- * @throws
- * - if the given address is not a valid address.
- * - if the function doesn't exist in the contract to call.
- *
- */
-export function call(
-  at: Address,
-  functionName: string,
-  args: Args,
-  coins: u64,
-): StaticArray<u8> {
-  return env.call(at.toString(), functionName, args.serialize(), coins);
-}
-
-/**
- * Calls a function from a remote contract in the current context.
- *
- * @remarks
- * Arguments serialization is to be handled by the caller and the callee.
- *
- * @param at - The address of the contract from where the function is located.
- * @param functionName - The name of the function to call in the current context.
- * @param args - The arguments of the function we are calling.
- *
- * @returns The return value of the executed function, serialized as a 'StaticArray<u8>'.
- *
- * @throws
- * - if the given address is not a valid address.
- * - if the function doesn't exist in the contract to call.
- *
- */
-export function localCall(
-  at: Address,
-  functionName: string,
-  args: Args,
-): StaticArray<u8> {
-  return env.localCall(at.toString(), functionName, args.serialize());
-}
-
-/**
- * Calls a function from a contract's bytecode in the current context.
- *
- * @remarks
- * This can be useful for testing or debugging purposes,
- * or for calling functions that are not meant to be called from outside the contract.
- * Arguments serialization is to be handled by the caller and the callee.
- *
- * @param bytecode - The bytecode of the contract containing the function to execute.
- * @param functionName - The name of the function to call in that contract.
- * @param args - The arguments of the function we are calling.
- *
- * @returns The return value of the executed function, serialized as a 'StaticArray<u8>'.
- *
- * @throws
- * - if the function doesn't exist in the bytecode
- *
- */
-export function localExecution(
-  bytecode: StaticArray<u8>,
-  functionName: string,
-  args: Args,
-): StaticArray<u8> {
-  return env.localExecution(bytecode, functionName, args.serialize());
-}
-
-/**
- * Retrieves the bytecode of the contract that is currently being executed.
- *
- * @remarks
- * Bytecode is a low-level representation of a smart contract's code that can be executed by the blockchain.
- *
- * @returns The bytecode of the contract, serialized as a 'StaticArray<u8>'.
- *
- */
-export function getBytecode(): StaticArray<u8> {
-  return env.getBytecode();
-}
-
-/**
- * Retrieves the bytecode of the remote contract at the given 'address'.
- *
- * @remarks
- * Bytecode is a low-level representation of a smart contract's code that can be executed by the blockchain.
- *
- * @param address - The address of the contract's bytecode to retrieve.
- *
- * @returns The bytecode of the contract, serialized as a 'StaticArray<u8>'.
- *
- * @throws
- * - if the given address is not a valid smart contract address.
- *
- */
-export function getBytecodeOf(address: Address): StaticArray<u8> {
-  return env.getBytecodeOf(address.toString());
-}
-
-/**
- * Determine if the caller has write access to the data stored in the called smart contract.
- *
- * @remarks
- * This function returns true exclusively when a new smart contract is deployed using the
- * 'create_new_sc_address()' function.
- * When calling {@link createSC}, the User or smart contract will be granted write
- * access to the created SC, but this privilege is limited to the context of this specific operation.
- *
- * @returns Returns true if the caller has write access; false otherwise.
- *
- */
-export function callerHasWriteAccess(): bool {
-  return env.callerHasWriteAccess();
-}
-
-/**
- * Creates a new smart contract on the ledger using its bytecode representation.
- *
- * @remarks
- * After executing this function, you will have write access on the newly generated contract.
- *
- * @see {@link callerHasWriteAccess} for more information.
- *
- * @param bytecode - The byte code of the contract to create.
- *
- * @returns The address of the newly created smart contract on the ledger.
- *
- */
-export function createSC(bytecode: StaticArray<u8>): Address {
-  return new Address(env.createSC(bytecode));
-}
-
-/**
- * Checks if the given function exists in a smart contract at the given address.
- *
- * @param address - The address of the contract to search in.
- * @param func - The name of the function to search.
- *
- * @returns true if the function exists, false otherwise.
- *
- * @throws
- * - if the given address is not a valid smart contract address.
- *
- */
-export function functionExists(address: Address, func: string): bool {
-  return env.functionExists(address.toString(), func);
-}
-
-/**
- * Generates an event that is then emitted by the blockchain.
- *
- * @param event - The string event to emit.
- *
- */
-export function generateEvent(event: string): void {
-  env.generateEvent(event);
-}
-
-/**
- * Transfers coins from the current address 'to' a given address.
- *
- * @param to - the address to send coins to.
- * @param amount - value in the smallest unit.
- * @see [massa units standard](https://github.com/massalabs/massa-standards/blob/main/units.md)
- *
- * @throws
- * - if the given address is not a valid address.
- * - if the balance of the current address is insufficient to make the transaction.
- *
- */
-export function transferCoins(to: Address, amount: u64): void {
-  env.transferCoins(to.toString(), amount);
-}
-
-/**
- * Transfers coins 'from' an address 'to' another address.
- *
- * @remarks
- * The transfer can only be done if the caller has write access to the sender's address.
- * @see {@link callerHasWriteAccess}
- *
- * @param from - the sender address.
- * @param to - the address to send coins to.
- * @param amount - value in the smallest unit.
- * @see [massa units standard](https://github.com/massalabs/massa-standards/blob/main/units.md)
- *
- * @throws
- * - if the sender's or the receiver address is not a valid address.
- * - if the balance of the sender's address is insufficient to make the transaction.
- *
- */
-export function transferCoinsOf(from: Address, to: Address, amount: u64): void {
-  env.transferCoinsOf(from.toString(), to.toString(), amount);
-}
-
-/**
- * Returns the balance of the current address.
- *
- * @returns - value in the smallest unit.
- *
- */
-export function balance(): u64 {
-  return env.balance();
-}
-
-/**
- * Returns the balance of the specified address.
- *
- * @param address - The address for which the balance is retrieved.
- *
- * @returns - value in the smallest unit.
- *
- * @throws
- * - if the given address is not a valid address.
- *
- */
-export function balanceOf(address: string): u64 {
-  return env.balanceOf(address);
-}
-
-/**
- * Checks if a given serialized 'key' is present in the operation datastore.
- *
- * @param key - the serialized key to look for in the datastore.
- *
- * @returns - true if key is present in datastore, false otherwise.
- *
- */
-export function hasOpKey(key: StaticArray<u8>): bool {
-  let result = env.hasOpKey(key);
-  // From https://doc.rust-lang.org/reference/types/boolean.html &&
-  // https://www.assemblyscript.org/types.html
-  // we can safely cast from u8 to bool
-  return bool(result[0]);
-}
-
-/**
- * Retrieves the data associated with the given key from the operation datastore.
- *
- * @param key - the serialized key to look for in the datastore.
- *
- * @returns - the serialized data associated with the given key as a byte array.
- *
- * @throws
- * - if the key is not present in the datastore.
- *
- */
-export function getOpData(key: StaticArray<u8>): StaticArray<u8> {
-  return env.getOpData(key);
-}
-
-/**
- * Retrieves all the keys from the operation datastore.
- *
- * @returns - a list of keys (e.g. a list of byte array)
- *
- */
-export function getOpKeys(): Array<StaticArray<u8>> {
-  let keysSer = env.getOpKeys();
-  return derKeys(keysSer);
-}
-
-/**
- * Retrieves all the keys from the operation datastore.
- *
- * @param prefix - the serialized prefix to filter the keys (optional)
- *
- * @returns - a list of keys (e.g. a list of byte array)
- *
- */
-export function getKeys(
-  prefix: StaticArray<u8> = new StaticArray<u8>(0),
-): Array<StaticArray<u8>> {
-  let keysSer = env.getKeys(prefix);
-  return derKeys(keysSer);
-}
-
-/**
- * Retrieves all the keys from the operation datastore from a remote address.
- *
- * @param address - the address in the datastore
- * @param prefix - the prefix to filter the keys (optional)
- *
- * @returns - a list of key (e.g. a list of byte array)
- *
- */
-export function getKeysOf(
-  address: string,
-  prefix: StaticArray<u8> = new StaticArray<u8>(0),
-): Array<StaticArray<u8>> {
-  let keysSer = env.getKeysOf(address, prefix);
-  return derKeys(keysSer);
-}
-
-/**
- * Retrieves the number of keys from serialized keys array
- *
- * @param arr - Uint8Array keys array
- *
- * @returns The number of keys in the keys array
- *
- */
-function getNumberOfKeys(keysSer: StaticArray<u8>): u32 {
-  // The first 4 bytes of the input array represent the number of keys
-  let arr = new Uint8Array(4);
-  arr[0] = keysSer[0];
-  arr[1] = keysSer[1];
-  arr[2] = keysSer[2];
-  arr[3] = keysSer[3];
-  let dv = new DataView(arr.buffer, arr.byteOffset, arr.byteLength);
-  let entryCount = dv.getUint32(0, true /* littleEndian */);
-  return entryCount;
-}
-
-/**
- * Deserializes an array of keys from the specified serialized format.
- *
- * @param keysSer - The serialized keys.
- *
- * @returns The deserialized keys.
- *
- * @remarks
- * 'keysSer' is a encoded array of keys. Each elements are encoded as follows:
- * ```text
- *|---------------|-----------|---------------|-----------------------------------------|
- *| Field         | Type      | Size in Bytes | Description                             |
- *|---------------|-----------|---------------|-----------------------------------------|
- *| L             | u32       | 4             | Total number of keys in the sequence.   |
- *| Key1_L        | u8        | 1             | Length of the Key number 1.             |
- *| Key1          | u8[Key1_L]| Variable      | The Key number 1.                       |
- *| Key2_L        | u8        | 1             | Length of the Key number 2.             |
- *| Key2          | u8[Key2_L]| Variable      | The Key number 2.                       |
- *| ...           |           |               | Data for additional keys (if any).      |
- *|---------------|-----------|---------------|-----------------------------------------|
- * ```
- * Then it returns an array of keys:
- * ```text
- * keys = [Key1, Key2, ...]
- * ```
- *
- */
-export function derKeys(keysSer: StaticArray<u8>): Array<StaticArray<u8>> {
-  if (keysSer.length == 0) return [];
-
-  const keyCount: u32 = getNumberOfKeys(keysSer);
-  const keys = new Array<StaticArray<u8>>(keyCount);
-
-  let cursor = 4;
-
-  for (let i: u32 = 0; i < keyCount; i++) {
-    const keyLen = keysSer[cursor];
-    const start = cursor + 1;
-    const end = start + keyLen;
-
-    keys[i] = keysSer.slice<StaticArray<u8>>(start, end);
-
-    cursor = end;
-  }
-  return keys;
-}
-
-/**
- * Converts the given string data into a base58 formatted string.
- *
- * @param data - the string data to convert to base58.
- *
- * @returns the converted data as a string.
- *
- */
-export function toBase58(data: string): string {
-  ERROR('toBase58 is not implemented yet');
-}
-
-/**
- * Checks if a signature is valid.
- *
- * @param publicKey - base58check encoded public key.
- * @param digest - digest message.
- * @param signature - base58check encoded signature.
- *
- * @returns 'true' if the signature is valid, 'false' otherwise.
- *
- */
-export function isSignatureValid(
-  publicKey: string,
-  digest: string,
-  signature: string,
-): bool {
-  return env.isSignatureValid(digest, signature, publicKey);
-}
-
-/**
- * Retrieves an 'Address' object from the given public key.
- *
- * @param pubKey - Base58check encoded public key of the address
- *
- * @returns the fetched address as an 'Address' object.
- *
- * @throws
- * - if the public key is invalid
- *
- */
-export function publicKeyToAddress(pubKey: string): Address {
-  return new Address(env.publicKeyToAddress(pubKey));
-}
-
-/**
- * Generates an unsafe random integer.
- *
- * @remarks
- * This function is unsafe because the random draws are predictable.
- *
- * @returns a random number.
- *
- */
-export function unsafeRandom(): i64 {
-  return env.unsafeRandom();
-}
-
-/**
- * Ask to schedule the execution of a function at a given address in the future.
- *
- * @remarks
- * The goal of sendMessage functionality is to send a message in the future, that will be executed as soon as possible
- * after the start period but not after the end period.
- *
- * You might want to use the sendMessage functionality:
- * - Having a smart contract called periodically, without a centralized bot;
- * - Having a smart contract that will trigger on the change of value (for example a change in price), of an other one;
- * - Having an object that evolves on the blockchain itself.
- *
- * This message allows you to make executions in the future and
- * they are executed deterministically on all nodes. The execution is made "as soon as possible" because there is a
- * priority on messages and a limit of messages possibly executed on each slot. More precisely, if you send a low amount
- * of `rawFee` then your message could possibly not executed directly at the first slot of the slot period.
- * If all of the slots of the specified period have a large load of messages, with more fees than you specified, then
- * it's highly likely that your message will never be executed.
- *
- * Additionally, there is an optional filter on a message that adds a new condition on the trigger instead of:
- * "as soon as possible in this range", it becomes
- * "as soon as possible, after this field has been updated, in the state, in this range".
- *
- * As a parameter, you can pass the `filterAddress`, and also an optional datastore `filterKey`.
- *
- * If you pass only an address, then the message will be executed only after:
- * "we are in the range, and a change has happened during this range on the `filterAddress`" (possibly balance etc).
- * If you provide a `filterKey`, the condition of the execution of the message is:
- * "we are in the range, and a change has happened during this range on the `filterAddress` at that datastore
- * `filterKey`"
- *
- * Note: serialization is to be handled at the caller and the callee level.
- *
- * @param at - Address of the contract
- * @param functionName - name of the function in that contract
- * @param validityStartPeriod - Period of the validity start slot
- * @param validityStartThread - Thread of the validity start slot
- * @param validityEndPeriod - Period of the validity end slot
- * @param validityEndThread - Thread of the validity end slot
- * @param maxGas - Maximum gas for the message execution
- * @param rawFee - Fee to be paid for message execution
- * @param coins - Coins of the sender
- * @param msg - function argument serialized in bytes
- * @param filterAddress - If you want your message to be trigger only
- * if a modification is made on a specific address precise it here
- * @param filterKey - If you want your message to be trigger only
- * if a modification is made on a specific storage key of the `filterAddress` precise it here
- *
- */
-export function sendMessage(
-  at: Address,
-  functionName: string,
-  validityStartPeriod: u64,
-  validityStartThread: u8,
-  validityEndPeriod: u64,
-  validityEndThread: u8,
-  maxGas: u64,
-  rawFee: u64,
-  coins: u64,
-  msg: StaticArray<u8>,
-  filterAddress: Address = new Address(),
-  filterKey: StaticArray<u8> = new StaticArray<u8>(0),
-): void {
-  env.sendMessage(
-    at.toString(),
-    functionName,
-    validityStartPeriod,
-    validityStartThread,
-    validityEndPeriod,
-    validityEndThread,
-    maxGas,
-    rawFee,
-    coins,
-    msg,
-    filterAddress.toString(),
-    filterKey,
-  );
-}
-
-/**
- * Convert the given file content to byteArray.
- *
- * @remarks
- * This function is dynamically replaced using the byteArray converter.
- * @see [as-transformer](https://github.com/massalabs/as/tree/main/packages/as-transformer#file2bytearray)
- *
- * @param filePath - the file path to convert
- *
- */
-export function fileToByteArray(
-  filePath: string, // eslint-disable-line @typescript-eslint/no-unused-vars
-): StaticArray<u8> {
-  abort('Please use transformer to dynamically include the file.');
-  return [];
-}
-
-/**
- * Retrieves the current period of the network.
- *
- * @returns the current period.
- *
- */
-export function currentPeriod(): u64 {
-  return env.currentPeriod();
-}
-
-/**
- * Retrieves the current thread of the execution context.
- *
- * @returns the current thread.
- *
- */
-export function currentThread(): u8 {
-  return env.currentThread();
-}
-
-/**
- * Constructs a pretty formatted event with given key and arguments.
- *
- * @remarks
- * The result is meant to be used with the {@link generateEvent} function.
- * It is useful to generate events from an array.
- *
- * @param key - the string event key.
- *
- * @param args - the string array arguments.
- *
- * @returns the stringified event.
- *
- */
-export function createEvent(key: string, args: Array<string>): string {
-  return `${key}:`.concat(args.join(','));
-}
-
-/**
- * Computes the SHA256 hash of the given `data`.
- *
- * @remarks
- * The SHA256 hash algorithm produces a 32-byte hash, which is returned as a `StaticArray<u8>`.
- *
- * @param data - The data to hash.
- *
- * @returns The SHA256 hash of the `data`, serialized as a `StaticArray<u8>`.
- *
- */
-export function sha256(data: StaticArray<u8>): StaticArray<u8> {
-  return env.sha256(data);
-}
-
-/**
- * Checks if the given address is valid.
- *
- * @param address - the string address to validate.
- *
- * @returns 'true' if the address is valid, 'false' otherwise.
- *
- */
-export function validateAddress(address: string): bool {
-  return env.validateAddress(address);
-}
+export * from './op-datastore';
+export * from './contract';
+export * from './utils';
+export * from './coins';
