+++ assembly/std/op-datastore/index.ts
@@ -1,0 +1,2 @@
+export * from './op-datastore';
+export * from './util';
